backend/requirements.txt
# Core requirements
Django==5.2.1
djangorestframework==3.15.1

# Document processing
pdfplumber==0.10.4
python-docx==1.1.0
PyMuPDF==1.24.1  # Alternative PDF processor

# Vector database & embeddings
chromadb==0.4.24
sentence-transformers==2.7.0
torch==2.2.2  # Required by sentence-transformers
numpy==1.26.4  # Required by various ML libraries

# LLM integration
groq==0.9.0
requests==2.31.0  # For API calls to LM Studio

# File handling
python-magic==0.4.27  # For better file type detection
filetype==1.2.0

# Utilities
tqdm==4.66.2  # Progress bars
loguru==0.7.2  # Enhanced logging (alternative to built-in logging)
python-dotenv==1.0.1  # For environment variables

# Development tools
ipython==8.22.2  # Better interactive console
pytest==8.1.1  # Testing framework